Windows: Factor out CloseExecFrom
