+++ libcontainer/utils/utils.go
@@ -4,9 +4,7 @@ import (
 	"crypto/rand"
 	"encoding/hex"
 	"io"
-	"io/ioutil"
 	"path/filepath"
-	"strconv"
 	"syscall"
 )
 
@@ -35,30 +33,6 @@ func ResolveRootfs(uncleanRootfs string) (string, error) {
 		return "", err
 	}
 	return filepath.EvalSymlinks(rootfs)
-}
-
-func CloseExecFrom(minFd int) error {
-	fdList, err := ioutil.ReadDir("/proc/self/fd")
-	if err != nil {
-		return err
-	}
-	for _, fi := range fdList {
-		fd, err := strconv.Atoi(fi.Name())
-		if err != nil {
-			// ignore non-numeric file names
-			continue
-		}
-
-		if fd < minFd {
-			// ignore descriptors lower than our specified minimum
-			continue
-		}
-
-		// intentionally ignore errors from syscall.CloseOnExec
-		syscall.CloseOnExec(fd)
-		// the cases where this might fail are basically file descriptors that have already been closed (including and especially the one that was created when ioutil.ReadDir did the "opendir" syscall)
-	}
-	return nil
 }
 
 // ExitStatus returns the correct exit status for a process based on if it
